@@ -145,7 +145,7 @@ namespace Veldrid.SPIRV
                         }
                     }
 
-                    SpirvReflection reflection = new SpirvReflection(
+                    SpirvReflection reflection = new(
                         vertexElements,
                         layouts);
 
@@ -248,7 +248,7 @@ namespace Veldrid.SPIRV
                         }
                     }
 
-                    SpirvReflection reflection = new SpirvReflection(
+                    SpirvReflection reflection = new(
                         Array.Empty<VertexElementDescription>(),
                         layouts);
 
@@ -286,7 +286,7 @@ namespace Veldrid.SPIRV
             }
 
             int macroCount = options.Macros.Length;
-            NativeMacroDefinition* macros = stackalloc NativeMacroDefinition[(int)macroCount];
+            NativeMacroDefinition* macros = stackalloc NativeMacroDefinition[macroCount];
             for (int i = 0; i < macroCount; i++)
             {
                 macros[i] = new NativeMacroDefinition(options.Macros[i]);
@@ -359,16 +359,16 @@ namespace Veldrid.SPIRV
 
         private static ShadercShaderKind GetShadercKind(ShaderStages stage)
         {
-            switch (stage)
+            return stage switch
             {
-                case ShaderStages.Vertex: return ShadercShaderKind.Vertex;
-                case ShaderStages.Geometry: return ShadercShaderKind.Geometry;
-                case ShaderStages.TessellationControl: return ShadercShaderKind.TessellationControl;
-                case ShaderStages.TessellationEvaluation: return ShadercShaderKind.TessellationEvaluation;
-                case ShaderStages.Fragment: return ShadercShaderKind.Fragment;
-                case ShaderStages.Compute: return ShadercShaderKind.Compute;
-                default: throw new SpirvCompilationException($"Invalid shader stage: {stage}");
-            }
+                ShaderStages.Vertex => ShadercShaderKind.Vertex,
+                ShaderStages.Geometry => ShadercShaderKind.Geometry,
+                ShaderStages.TessellationControl => ShadercShaderKind.TessellationControl,
+                ShaderStages.TessellationEvaluation => ShadercShaderKind.TessellationEvaluation,
+                ShaderStages.Fragment => ShadercShaderKind.Fragment,
+                ShaderStages.Compute => ShadercShaderKind.Compute,
+                _ => throw new SpirvCompilationException($"Invalid shader stage: {stage}"),
+            };
         }
     }
 }
