@@ -1,4 +1,5 @@
-﻿using System.Text;
+﻿using System.Diagnostics.CodeAnalysis;
+using System.Text;
 
 namespace Veldrid.SPIRV
 {
@@ -53,8 +54,8 @@ namespace Veldrid.SPIRV
 
                 return new Shader[]
                 {
-                    factory.CreateShader(in vertexShaderDescription),
-                    factory.CreateShader(in fragmentShaderDescription)
+                    factory.CreateShader(vertexShaderDescription),
+                    factory.CreateShader(fragmentShaderDescription)
                 };
             }
 
@@ -122,7 +123,7 @@ namespace Veldrid.SPIRV
             if (backend == GraphicsBackend.Vulkan)
             {
                 computeShaderDescription.ShaderBytes = EnsureSpirv(computeShaderDescription);
-                return factory.CreateShader(in computeShaderDescription);
+                return factory.CreateShader(computeShaderDescription);
             }
 
             CrossCompileTarget target = GetCompilationTarget(factory.BackendType);
@@ -164,6 +165,7 @@ namespace Veldrid.SPIRV
             }
         }
 
+        [SuppressMessage("Style", "IDE0066:Convert switch statement to expression", Justification = "<Pending>")]
         private static byte[] GetBytes(GraphicsBackend backend, string code)
         {
             switch (backend)
@@ -173,7 +175,7 @@ namespace Veldrid.SPIRV
                 case GraphicsBackend.OpenGLES:
                     return Encoding.ASCII.GetBytes(code);
                 case GraphicsBackend.Metal:
-                    return Encoding.UTF8.GetBytes(code);
+                    return Util.UTF8.GetBytes(code);
                 default:
                     throw new SpirvCompilationException($"Invalid GraphicsBackend: {backend}");
             }
@@ -181,19 +183,14 @@ namespace Veldrid.SPIRV
 
         private static CrossCompileTarget GetCompilationTarget(GraphicsBackend backend)
         {
-            switch (backend)
+            return backend switch
             {
-                case GraphicsBackend.Direct3D11:
-                    return CrossCompileTarget.HLSL;
-                case GraphicsBackend.OpenGL:
-                    return CrossCompileTarget.GLSL;
-                case GraphicsBackend.Metal:
-                    return CrossCompileTarget.MSL;
-                case GraphicsBackend.OpenGLES:
-                    return CrossCompileTarget.ESSL;
-                default:
-                    throw new SpirvCompilationException($"Invalid GraphicsBackend: {backend}");
-            }
+                GraphicsBackend.Direct3D11 => CrossCompileTarget.HLSL,
+                GraphicsBackend.OpenGL => CrossCompileTarget.GLSL,
+                GraphicsBackend.Metal => CrossCompileTarget.MSL,
+                GraphicsBackend.OpenGLES => CrossCompileTarget.ESSL,
+                _ => throw new SpirvCompilationException($"Invalid GraphicsBackend: {backend}"),
+            };
         }
     }
 }
